+++ tests/auto/tools/qmakelib/parsertest.cpp
@@ -42,7 +42,7 @@ public:
     TokenStream &operator<<(uint n) { ts += QChar(n & 0xffff); ts += QChar(n >> 16); return *this; }
     TokenStream &operator<<(QStringView s) { ts += s; return *this; }
     TokenStream &operator<<(const ProString &s) { return *this << ushort(s.length()) << s.toQStringView(); }
-    TokenStream &operator<<(const ProKey &s) { return *this << s.hash() << s.toString(); }
+    TokenStream &operator<<(const ProKey &s) { return *this << uint(s.hash()) << s.toString(); }
 
 private:
     QString ts;
